+++ src/main/java/cn/nukkit/block/BlockID.java
@@ -1,5 +1,8 @@
 package cn.nukkit.block;
 
+import cn.nukkit.api.PowerNukkitOnly;
+import cn.nukkit.api.Since;
+
 public interface BlockID {
     int AIR = 0;
     int STONE = 1;
@@ -71,6 +74,7 @@ public interface BlockID {
     int TORCH = 50;
     int FIRE = 51;
     int MONSTER_SPAWNER = 52;
+    @PowerNukkitOnly @Since("1.3.2.0-PN") int MOB_SPAWNER = MONSTER_SPAWNER;
     int WOOD_STAIRS = 53;
     int WOODEN_STAIRS = 53;
     int OAK_WOOD_STAIRS = 53;