+++ src/main/java/cn/nukkit/item/Item.java
@@ -548,6 +548,12 @@ public class Item implements Cloneable, BlockID, ItemID {
     }
 
     public static Item get(int id, Integer meta, int count, byte[] tags) {
+        return getItem(id, meta, count, tags, true);
+    }
+
+    @PowerNukkitOnly("Created with a new name to avoid possible signature conflicts in future")
+    @Since("1.3.2.0-PN")
+    public static Item getItem(int id, Integer meta, int count, byte[] tags, boolean selfUpgrade) {
         try {
             Class c = null;
             if (id < 0) {
@@ -573,8 +579,12 @@ public class Item implements Cloneable, BlockID, ItemID {
             if (tags.length != 0) {
                 item.setCompoundTag(tags);
             }
-
-            return item.selfUpgrade();
+            
+            if (selfUpgrade) {
+                item = item.selfUpgrade();
+            }
+            
+            return item;
         } catch (Exception e) {
             return new Item(id, meta, count).setCompoundTag(tags);
         }