+++ src/main/java/cn/nukkit/item/ItemID.java
@@ -290,7 +290,6 @@ public interface ItemID {
     
     //int SOUL_CAMPFIRE = 801;
     
-    //TODO: Make all this work!
     @Since("1.3.2.0-PN") @PowerNukkitOnly int COD_BUCKET = 802;
     @Since("1.3.2.0-PN") @PowerNukkitOnly int GHAST_SPAWN_EGG = 803;
     @Since("1.3.2.0-PN") @PowerNukkitOnly int FLOWER_BANNER_PATTERN = 804;