+++ src/main/java/cn/nukkit/level/Level.java
@@ -23,6 +23,7 @@ import cn.nukkit.event.weather.LightningStrikeEvent;
 import cn.nukkit.item.Item;
 import cn.nukkit.item.ItemBlock;
 import cn.nukkit.item.ItemBucket;
+import cn.nukkit.item.ItemSpawnEgg;
 import cn.nukkit.item.enchantment.Enchantment;
 import cn.nukkit.level.biome.Biome;
 import cn.nukkit.level.format.Chunk;
@@ -2256,6 +2257,20 @@ public class Level implements ChunkManager, Metadatable {
 
     @PowerNukkitDifference(info = "PowerNukkit#403", since = "1.3.1.2-PN")
     public Item useItemOn(Vector3 vector, Item item, BlockFace face, float fx, float fy, float fz, Player player, boolean playSound) {
+        // Hack for backward compatibility with MobPlugin
+        Item input;
+        if (item instanceof ItemSpawnEgg) {
+            input = ((ItemSpawnEgg) item).getLegacySpawnEgg();
+        } else {
+            input = item.clone();
+        }
+        Item output = useItemOn0(vector, input, face, fx, fy, fz, player, playSound);
+        item.setCount(input.getCount());
+        item.setCompoundTag(input.getCompoundTag());
+        return output;
+    }
+
+    private Item useItemOn0(Vector3 vector, Item item, BlockFace face, float fx, float fy, float fz, Player player, boolean playSound) {
         Block target = this.getBlock(vector);
         Block block = target.getSide(face);